+++ templates/client.mustache
@@ -17,8 +17,6 @@
 			<link href="styles/client.css" rel="stylesheet" type="text/css">
 
 		<!-- Scripts -->
-		<script src="scripts/lib/OBJFile.js"></script>
-		<script src="scripts/whenDocReady.js"></script>
 		<script src="scripts/dist/client.js"></script>
 	</head>
 	<body>
